frontend/requirements.txt
pyserial
numpy
PyQt5
pyqtgraph